+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
   - Livesession backend rewrite
   - Add sentry
   - Create a live from the website (#2134)
+  - Integrate webinar dashboard (#2135)
 - Add a License Manager widget for LTI VOD view
 - Add a title to the classroom file dropzone
 - Add can_edit property on a serialized video